Adding DefaultServer to reverse proxy config. (#21)
Without this, we rely on the ordering of server blocks within the generated
nginx config in order to determine where to route traffic without a host
header. Rather than leaving it to chance, we want to explicitly indicate
that a given server is the default server.

Note: this patch makes no attempt to prevent duplicate default_servers
listening on a given port. The behavior in that case is undefiend and has
not been tested. The only code path setting it explicitly at the moment is
via `DefaultHTTPReverseProxyServers`.