Use StaticCode when a location has no upstream servers. (#22)
* Don't render invalid upstream blocks.

When no servers are discovered for an upstream, we do not render it.

* Use StaticCode when a location has no upstream servers.

Farva cannot render valid upstream blocks when there are no servers discovered,
therefore we do not render upstreams without servers.

Since we do not render upstreams without servers, we cannot refer to these
invalid upstream blocks by name. We defend against this situation when reading
from the k8s API's by configuring the invalid httpReverseProxyLocation to
return a static 503 instead of proxying to it's invalid upstreams.